requirements.txt
discord.py==2.1.*
python-dotenv==0.21.*
openai==1.2.0
PyYAML==6.0
dacite==1.6.*
Spacy==3.0.*